+++ forui/lib/src/widgets/time_field/picker/picker_time_field.dart
@@ -91,7 +91,6 @@ class _PickerTimeFieldState extends _FTimeFieldState<_PickerTimeField> {
   void initState() {
     super.initState();
     _controller._picker.addListener(_updateTextController);
-    _controller.popover.addListener(_updateFocus);
     _controller.addValueListener(_onChange);
   }
 
@@ -116,13 +115,11 @@ class _PickerTimeFieldState extends _FTimeFieldState<_PickerTimeField> {
         _controller.dispose();
       } else {
         _controller._picker.removeListener(_updateTextController);
-        _controller.popover.removeListener(_updateFocus);
         _controller.removeValueListener(_onChange);
       }
 
       _controller = widget.controller ?? FTimeFieldController(vsync: this, initialTime: _controller.value);
       _controller._picker.addListener(_updateTextController);
-      _controller.popover.addListener(_updateFocus);
       _controller.addValueListener(_onChange);
       _updateTextController();
     }
@@ -149,12 +146,6 @@ class _PickerTimeFieldState extends _FTimeFieldState<_PickerTimeField> {
     });
   }
 
-  void _updateFocus() {
-    if (!_controller.popover.shown) {
-      _focus.unfocus();
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     final style = widget.style ?? context.theme.timeFieldStyle;
@@ -179,7 +170,8 @@ class _PickerTimeFieldState extends _FTimeFieldState<_PickerTimeField> {
             expands: widget.expands,
             mouseCursor: widget.mouseCursor,
             canRequestFocus: widget.canRequestFocus,
-            onTap: _controller.popover.toggle,
+            onTap: _onTap,
+            onTapAlwaysCalled: true,
             hint: widget.hint ?? localizations.dateFieldHint,
             readOnly: true,
             enableInteractiveSelection: false,
@@ -207,10 +199,20 @@ class _PickerTimeFieldState extends _FTimeFieldState<_PickerTimeField> {
                   style: style,
                   hour24: widget.hour24,
                   properties: widget,
-                  child: widget.builder(context, (style, styles.$1, styles.$2), child),
+                  autofocus: true,
+                  fieldFocusNode: _focus,
+                  child: CallbackShortcuts(
+                    bindings: {const SingleActivator(LogicalKeyboardKey.enter): _onTap},
+                    child: widget.builder(context, (style, styles.$1, styles.$2), child),
+                  ),
                 ),
           ),
     );
+  }
+
+  void _onTap() {
+    _controller.popover.shown ? _focus.requestFocus() : _focus.unfocus();
+    _controller.popover.toggle();
   }
 
   @override
@@ -219,7 +221,6 @@ class _PickerTimeFieldState extends _FTimeFieldState<_PickerTimeField> {
       _controller.dispose();
     } else {
       _controller._picker.removeListener(_updateTextController);
-      _controller.popover.removeListener(_updateFocus);
       _controller.removeValueListener(_onChange);
     }
 
@@ -236,6 +237,8 @@ class _PickerPopover extends StatelessWidget {
   final FTimeFieldStyle style;
   final FTimeFieldPickerProperties properties;
   final bool hour24;
+  final bool autofocus;
+  final FocusNode? fieldFocusNode;
   final Widget child;
 
   const _PickerPopover({
@@ -243,6 +246,8 @@ class _PickerPopover extends StatelessWidget {
     required this.style,
     required this.properties,
     required this.hour24,
+    required this.autofocus,
+    required this.fieldFocusNode,
     required this.child,
   });
 
@@ -257,6 +262,8 @@ class _PickerPopover extends StatelessWidget {
     shift: properties.shift,
     offset: properties.offset,
     hideOnTapOutside: properties.hideOnTapOutside,
+    autofocus: autofocus,
+    shortcuts: {const SingleActivator(LogicalKeyboardKey.escape): _hide},
     popoverBuilder:
         (_, _, _) => TextFieldTapRegion(
           child: Padding(
@@ -273,6 +280,11 @@ class _PickerPopover extends StatelessWidget {
     child: child,
   );
 
+  void _hide() {
+    fieldFocusNode?.requestFocus();
+    controller.popover.hide();
+  }
+
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
@@ -280,6 +292,8 @@ class _PickerPopover extends StatelessWidget {
       ..add(DiagnosticsProperty('controller', controller))
       ..add(DiagnosticsProperty('style', style))
       ..add(DiagnosticsProperty('properties', this.properties))
-      ..add(DiagnosticsProperty('hour24', hour24));
+      ..add(DiagnosticsProperty('hour24', hour24))
+      ..add(FlagProperty('autofocus', value: autofocus, ifTrue: 'autofocus'))
+      ..add(DiagnosticsProperty('fieldFocusNode', fieldFocusNode));
   }
 }