+++ forui/test/src/widgets/date_field/calendar/calendar_date_field_golden_test.dart
@@ -1,5 +1,5 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 import 'package:flutter_test/flutter_test.dart';
 
@@ -144,42 +144,22 @@ void main() {
       await expectLater(find.byType(TestScaffold), matchesGoldenFile('date-field/${theme.name}/calendar/error.png'));
     });
 
-    testWidgets('${theme.name} tap outside unfocuses on Android/iOS', (tester) async {
+    testWidgets('${theme.name} keyboard navigation', (tester) async {
       await tester.pumpWidget(
-        TestScaffold.app(theme: theme.data, alignment: Alignment.topCenter, child: FDateField.calendar(key: key)),
+        TestScaffold.app(theme: theme.data, alignment: Alignment.topCenter, child: FDateField.calendar()),
       );
 
-      await tester.tap(find.byKey(key));
+      await tester.sendKeyEvent(LogicalKeyboardKey.tab);
+      await tester.sendKeyEvent(LogicalKeyboardKey.enter);
       await tester.pumpAndSettle();
 
-      await tester.tapAt(Offset.zero);
+      await tester.sendKeyEvent(LogicalKeyboardKey.tab);
       await tester.pumpAndSettle();
 
       await expectLater(
         find.byType(TestScaffold),
-        matchesGoldenFile('date-field/${theme.name}/calendar/mobile-unfocused.png'),
+        matchesGoldenFile('date-field/${theme.name}/calendar/keyboard-navigation.png'),
       );
-    });
-
-    testWidgets('${theme.name} tap outside unfocuses on desktop', (tester) async {
-      debugDefaultTargetPlatformOverride = TargetPlatform.macOS;
-
-      await tester.pumpWidget(
-        TestScaffold.app(theme: theme.data, alignment: Alignment.topCenter, child: FDateField.calendar(key: key)),
-      );
-
-      await tester.tap(find.byKey(key));
-      await tester.pumpAndSettle();
-
-      await tester.tapAt(const Offset(500, 500));
-      await tester.pumpAndSettle();
-
-      await expectLater(
-        find.byType(TestScaffold),
-        matchesGoldenFile('date-field/${theme.name}/calendar/desktop-unfocused.png'),
-      );
-
-      debugDefaultTargetPlatformOverride = null;
     });
   }
 }