+++ forui/test/src/widgets/time_field/picker/picker_time_field_golden_test.dart
@@ -1,5 +1,5 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 import 'package:flutter_test/flutter_test.dart';
 
@@ -114,42 +114,20 @@ void main() {
       await expectLater(find.byType(TestScaffold), matchesGoldenFile('time-field/${theme.name}/picker/error.png'));
     });
 
-    testWidgets('${theme.name} tap outside unfocuses on Android/iOS', (tester) async {
-      await tester.pumpWidget(
-        TestScaffold.app(theme: theme.data, alignment: Alignment.topCenter, child: const FTimeField.picker(key: key)),
-      );
+    testWidgets('${theme.name} keyboard navigation', (tester) async {
+      await tester.pumpWidget(TestScaffold.app(theme: theme.data, child: const FTimeField.picker(key: key)));
 
-      await tester.tap(find.byKey(key));
+      await tester.sendKeyEvent(LogicalKeyboardKey.tab);
+      await tester.sendKeyEvent(LogicalKeyboardKey.enter);
       await tester.pumpAndSettle();
 
-      await tester.tapAt(Offset.zero);
-      await tester.pumpAndSettle();
-
-      await expectLater(
-        find.byType(TestScaffold),
-        matchesGoldenFile('time-field/${theme.name}/picker/mobile-unfocused.png'),
-      );
-    });
-
-    testWidgets('${theme.name} tap outside unfocuses on desktop', (tester) async {
-      debugDefaultTargetPlatformOverride = TargetPlatform.macOS;
-
-      await tester.pumpWidget(
-        TestScaffold.app(theme: theme.data, alignment: Alignment.topCenter, child: const FTimeField.picker(key: key)),
-      );
-
-      await tester.tap(find.byKey(key));
-      await tester.pumpAndSettle();
-
-      await tester.tapAt(Offset.zero);
+      await tester.sendKeyEvent(LogicalKeyboardKey.tab);
       await tester.pumpAndSettle();
 
       await expectLater(
         find.byType(TestScaffold),
-        matchesGoldenFile('time-field/${theme.name}/picker/desktop-unfocused.png'),
+        matchesGoldenFile('time-field/${theme.name}/picker/keyboard-navigation.png'),
       );
-
-      debugDefaultTargetPlatformOverride = null;
     });
   }
 }